+++ History.md
@@ -3,7 +3,7 @@
   * Add file-instrumentation only goal `jscover-file:instrument` (https://github.com/tntim96/JSCover-maven-plugin/issues/14)
   * Add merge goal `jscover-report:merge` (https://github.com/tntim96/JSCover-maven-plugin/issues/15)
   * Add format goal `jscover-report:format` (https://github.com/tntim96/JSCover-maven-plugin/issues/15)
-  * JSCover 1.0.22
+  * JSCover 1.0.22, Upgrade Jasmine 2.3.4->2.4.1
 
 1.0.15 / 2015-08-26
 ==================